requirements.txt
aiopg==0.13.0
aiopyramid==0.3.6
appdirs==1.4.3
asyncpg==0.9.0
attrs==16.0.0
-e hg+ssh://dev/~/repos/campin@0a2e3b4a425076265fcf6d69bde57b0bf30a42e2#egg=campin
cffi==1.7.0
cryptography==1.4
cssselect==0.9.2
FormEncode==1.3.1
googlemaps==2.4.4
greenlet==0.4.12
hupper==0.4.4
idna==2.1
lxml==3.6.0
Mako==1.0.6
MarkupSafe==1.0
olefile==0.44
packaging==16.8
parsel==1.0.2
PasteDeploy==1.5.2
Pillow==4.0.0
psycopg2==2.7.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==16.0.0
pyparsing==2.2.0
pyramid==1.8.3
pyramid-debugtoolbar==3.0.5
pyramid-mako==1.0.2
pyramid-simpleform==0.6.1
python-dateutil==2.6.0
queuelib==1.4.2
repoze.lru==0.6
requests==2.10.0
Scrapy==1.1.0
service-identity==16.0.0
six==1.10.0
translationstring==1.3
Twisted==16.3.0
txpostgres==1.6.0
uWSGI==2.0.11.2
venusian==1.0
w3lib==1.14.2
WebHelpers==1.3
WebOb==1.7.2
zope.deprecation==4.2.0
zope.interface==4.2.0
